requirements.txt
asgiref==3.3.4
backcall==0.2.0
beautifulsoup4==4.9.3
certifi==2020.12.5
chardet==4.0.0
colorama==0.4.4
decorator==5.0.7
Django==3.2.3
django-appconf==1.0.4
django-bootstrap-v5==1.0.2
django-contrib-postgres==0.0.1
django-cors-headers==3.7.0
django-extensions==3.1.3
django-imagekit==4.0.2
django-seed==0.2.2
djangorestframework==3.12.4
djangorestframework-jwt==1.11.0
Faker==8.1.1
idna==2.10
importlib-metadata==2.1.1
ipython==7.22.0
ipython-genutils==0.2.0
jedi==0.18.0
parso==0.8.2
pickleshare==0.7.5
pilkit==2.0
prompt-toolkit==3.0.18
psycopg2==2.8.6
Pygments==2.8.1
PyJWT==1.7.1
python-dateutil==2.8.1
pytz==2021.1
requests==2.25.1
six==1.15.0
soupsieve==2.2.1
sqlparse==0.4.1
text-unidecode==1.3
traitlets==5.0.5
urllib3==1.26.3
wcwidth==0.2.5
zipp==3.4.1
